+++ src/misc/get-file-info.ts
@@ -1,4 +1,3 @@
-import * as fs from 'fs';
 import * as crypto from 'crypto';
 import * as stream from 'stream';
 import * as util from 'util';
@@ -7,6 +6,11 @@ import isSvg from 'is-svg';
 import * as probeImageSize from 'probe-image-size';
 import * as sharp from 'sharp';
 import { encode } from 'blurhash';
+import { awaitAll } from '@/prelude/await-all';
+import { preventEmptyStream } from './stream/prevent-empty';
+import { readableRead } from './stream/read';
+import { createReadStream } from 'fs';
+import { cloneStream } from './stream/clone';
 
 const pipeline = util.promisify(stream.pipeline);
 
@@ -33,26 +37,42 @@ const TYPE_SVG = {
 	ext: 'svg'
 };
 
+export function getFileInfoByPath(path: string): Promise<FileInfo> {
+	return getFileInfo(createReadStream(path));
+}
+
 /**
  * Get file information
  */
-export async function getFileInfo(path: string): Promise<FileInfo> {
+export async function getFileInfo(readable: stream.Readable): Promise<FileInfo> {
 	const warnings = [] as string[];
 
-	const size = await getFileSize(path);
-	const md5 = await calcHash(path);
+	const streamCopy = cloneStream(readable);
 
-	let type = await detectType(path);
+	readableRead(readable);
+
+	let type = await detectType(readable);
+
+	const sizeDetectable = ['image/jpeg', 'image/gif', 'image/png', 'image/apng', 'image/webp', 'image/bmp', 'image/tiff', 'image/svg+xml', 'image/vnd.adobe.photoshop'].includes(type.mime);
+	const blurhashEnabled = ['image/jpeg', 'image/gif', 'image/png', 'image/apng', 'image/webp', 'image/svg+xml'].includes(type.mime);
+
+	const md5Promise = calcHash(streamCopy);
+	const sizePromise = getFileSize(streamCopy);
+	const imageSizePromise = sizeDetectable ? detectImageSize(streamCopy).catch(e => {
+		warnings.push(`detectImageSize failed: ${e}`);
+		return undefined;
+	}) : Promise.resolve(undefined);
+	const blurhashPromise = blurhashEnabled ? getBlurhash(streamCopy).catch(e => {
+		warnings.push(`getBlurhash failed: ${e}`);
+		return undefined;
+	}) : Promise.resolve(undefined);
 
 	// image dimensions
 	let width: number | undefined;
 	let height: number | undefined;
 
-	if (['image/jpeg', 'image/gif', 'image/png', 'image/apng', 'image/webp', 'image/bmp', 'image/tiff', 'image/svg+xml', 'image/vnd.adobe.photoshop'].includes(type.mime)) {
-		const imageSize = await detectImageSize(path).catch(e => {
-			warnings.push(`detectImageSize failed: ${e}`);
-			return undefined;
-		});
+	if (sizeDetectable) {
+		const imageSize = await imageSizePromise;
 
 		// うまく判定できない画像は octet-stream にする
 		if (!imageSize) {
@@ -72,42 +92,47 @@ export async function getFileInfo(path: string): Promise<FileInfo> {
 		}
 	}
 
-	let blurhash: string | undefined;
-
-	if (['image/jpeg', 'image/gif', 'image/png', 'image/apng', 'image/webp', 'image/svg+xml'].includes(type.mime)) {
-		blurhash = await getBlurhash(path).catch(e => {
-			warnings.push(`getBlurhash failed: ${e}`);
-			return undefined;
-		});
-	}
-
-	return {
-		size,
-		md5,
+	return awaitAll({
+		size: sizePromise,
+		md5: md5Promise,
 		type,
 		width,
 		height,
-		blurhash,
+		blurhash: blurhashPromise,
 		warnings,
-	};
+	});
 }
 
 /**
  * Detect MIME Type and extension
  */
-export async function detectType(path: string) {
-	// Check 0 byte
-	const fileSize = await getFileSize(path);
+export async function detectType(readable: stream.Readable) {
+	const chunks: Uint8Array[] = [];
+
+	const streamPromise = new Promise<Buffer | void>(res => {
+		cloneStream(readable)
+			.on('data', chunk => chunks.push(chunk))
+			.on('end', () => res(Buffer.concat(chunks)))
+			.on('error', e => res());
+	});
+
+	const fileSizePromise = getFileSize(readable);
+
+	const typePromise = fileType.fromStream(cloneStream(readable));
+
+	const [ fileSize, type ] = await Promise.all([fileSizePromise, typePromise]);
+
 	if (fileSize === 0) {
 		return TYPE_OCTET_STREAM;
 	}
 
-	const type = await fileType.fromFile(path);
-
 	if (type) {
 		// XMLはSVGかもしれない
-		if (type.mime === 'application/xml' && await checkSvg(path)) {
-			return TYPE_SVG;
+		if (type.mime === 'application/xml') {
+			const buffer = await streamPromise;
+			if (buffer && checkSvg(buffer)) {
+				return TYPE_SVG;
+			}
 		}
 
 		return {
@@ -117,7 +142,8 @@ export async function detectType(path: string) {
 	}
 
 	// 種類が不明でもSVGかもしれない
-	if (await checkSvg(path)) {
+	const buffer = await streamPromise;
+	if (buffer && checkSvg(buffer)) {
 		return TYPE_SVG;
 	}
 
@@ -128,11 +154,10 @@ export async function detectType(path: string) {
 /**
  * Check the file is SVG or not
  */
-export async function checkSvg(path: string) {
+export function checkSvg(buffer: Buffer) {
 	try {
-		const size = await getFileSize(path);
-		if (size > 1 * 1024 * 1024) return false;
-		return isSvg(fs.readFileSync(path));
+		if (buffer.length > 1 * 1024 * 1024) return false;
+		return isSvg(buffer);
 	} catch {
 		return false;
 	}
@@ -141,47 +166,50 @@ export async function checkSvg(path: string) {
 /**
  * Get file size
  */
-export async function getFileSize(path: string): Promise<number> {
-	const getStat = util.promisify(fs.stat);
-	return (await getStat(path)).size;
+export async function getFileSize(readable: stream.Readable): Promise<number> {
+	let length = 0;
+	for await (const chunk of cloneStream(readable)) {
+		length += chunk.length;
+	}
+	return length;
 }
 
 /**
  * Calculate MD5 hash
  */
-async function calcHash(path: string): Promise<string> {
+async function calcHash(readable: stream.Readable): Promise<string> {
 	const hash = crypto.createHash('md5').setEncoding('hex');
-	await pipeline(fs.createReadStream(path), hash);
+	await pipeline(cloneStream(readable), hash);
 	return hash.read();
 }
 
 /**
  * Detect dimensions of image
  */
-async function detectImageSize(path: string): Promise<{
+async function detectImageSize(readable: stream.Readable): Promise<{
 	width: number;
 	height: number;
 	wUnits: string;
 	hUnits: string;
 }> {
-	const readable = fs.createReadStream(path);
-	const imageSize = await probeImageSize(readable);
-	readable.destroy();
+	const streamCopy = cloneStream(readable);
+	const imageSize = await probeImageSize(streamCopy);
 	return imageSize;
 }
 
 /**
  * Calculate average color of image
  */
-function getBlurhash(path: string): Promise<string> {
+function getBlurhash(readable: stream.Readable): Promise<string> {
 	return new Promise((resolve, reject) => {
-		sharp(path)
+		const generator = sharp()
 			.raw()
 			.ensureAlpha()
 			.resize(64, 64, { fit: 'inside' })
-			.toBuffer((err, buffer, { width, height }) => {
+			.toBuffer((err, buffer, info) => {
 				if (err) return reject(err);
 
+				const { width, height } = info;
 				let hash;
 
 				try {
@@ -192,5 +220,7 @@ function getBlurhash(path: string): Promise<string> {
 
 				resolve(hash);
 			});
+
+		pipeline(readable, new stream.PassThrough(), preventEmptyStream(), generator).catch(reject);
 	});
 }