+++ src/queue/processors/db/export-following.ts
@@ -9,6 +9,7 @@ import { getFullApAccount } from '@/misc/convert-host';
 import { Users, Followings } from '@/models/index';
 import { MoreThan } from 'typeorm';
 import { DbUserJobData } from '@/queue/types';
+import { createTemp } from '@/misc/create-temp';
 
 const logger = queueLogger.createSubLogger('export-following');
 
@@ -22,12 +23,7 @@ export async function exportFollowing(job: Bull.Job<DbUserJobData>, done: any): 
 	}
 
 	// Create temp file
-	const [path, cleanup] = await new Promise<[string, any]>((res, rej) => {
-		tmp.file((e, path, fd, cleanup) => {
-			if (e) return rej(e);
-			res([path, cleanup]);
-		});
-	});
+	const [path, cleanup] = await createTemp();
 
 	logger.info(`Temp file is ${path}`);