+++ src/services/drive/generate-video-thumbnail.ts
@@ -1,15 +1,27 @@
 import * as fs from 'fs';
-import * as tmp from 'tmp';
-import { IImage, convertToJpeg } from './image-processor';
+import * as stream from 'stream';
+import * as util from 'util';
+import { IReadableImage, convertToJpeg } from './image-processor';
 import * as FFmpeg from 'fluent-ffmpeg';
+import { createTemp } from '@/misc/create-temp';
+import { readableRead } from '@/misc/stream/read';
 
-export async function GenerateVideoThumbnail(path: string): Promise<IImage> {
-	const [outDir, cleanup] = await new Promise<[string, any]>((res, rej) => {
-		tmp.dir((e, path, cleanup) => {
-			if (e) return rej(e);
-			res([path, cleanup]);
-		});
-	});
+const pipeline = util.promisify(stream.pipeline);
+
+export async function GenerateVideoThumbnailFromStream(readable: stream.Readable): Promise<IReadableImage> {
+	const [path, cleanupVideo] = await createTemp();
+
+	await pipeline(readable, fs.createWriteStream(path));
+
+	const thumbnail = await GenerateVideoThumbnail(path);
+
+	cleanupVideo();
+
+	return thumbnail;
+}
+
+export async function GenerateVideoThumbnail(path: string): Promise<IReadableImage> {
+	const [outDir, cleanupDir] = await createTemp();
 
 	await new Promise((res, rej) => {
 		FFmpeg({
@@ -27,11 +39,9 @@ export async function GenerateVideoThumbnail(path: string): Promise<IImage> {
 
 	const outPath = `${outDir}/output.png`;
 
-	const thumbnail = await convertToJpeg(outPath, 498, 280);
+	const thumbnail = convertToJpeg(readableRead(fs.createReadStream(outPath)), 498, 280);
 
-	// cleanup
-	await fs.promises.unlink(outPath);
-	cleanup();
+	cleanupDir();
 
 	return thumbnail;
 }