+++ src/services/drive/internal-storage.ts
@@ -3,6 +3,10 @@ import * as Path from 'path';
 import { fileURLToPath } from 'url';
 import { dirname } from 'path';
 import config from '@/config/index';
+import * as stream from 'stream';
+import * as util from 'util';
+
+const pipeline = util.promisify(stream.pipeline);
 
 //const _filename = fileURLToPath(import.meta.url);
 const _filename = __filename;
@@ -23,9 +27,9 @@ export class InternalStorage {
 		return `${config.url}/files/${key}`;
 	}
 
-	public static saveFromBuffer(key: string, data: Buffer) {
+	public static async saveFromStream(key: string, readable: stream.Readable) {
 		fs.mkdirSync(InternalStorage.path, { recursive: true });
-		fs.writeFileSync(InternalStorage.resolvePath(key), data);
+		await pipeline(readable, fs.createWriteStream(InternalStorage.resolvePath(key)));
 		return `${config.url}/files/${key}`;
 	}