+++ test/get-file-stream-info.ts
@@ -1,11 +1,12 @@
 import * as assert from 'assert';
 import { async } from './utils';
-import { getFileInfo } from '../src/misc/get-file-info';
+import { getFileInfoByPath } from '../src/misc/get-file-info';
+import { createReadStream } from 'fs';
 
-describe('Get file info', () => {
+describe('Get file info (stream)', () => {
 	it('Empty file', async (async () => {
 		const path = `${__dirname}/resources/emptyfile`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -22,7 +23,7 @@ describe('Get file info', () => {
 
 	it('Generic JPEG', async (async () => {
 		const path = `${__dirname}/resources/Lenna.jpg`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -39,7 +40,7 @@ describe('Get file info', () => {
 
 	it('Generic APNG', async (async () => {
 		const path = `${__dirname}/resources/anime.png`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -56,7 +57,7 @@ describe('Get file info', () => {
 
 	it('Generic AGIF', async (async () => {
 		const path = `${__dirname}/resources/anime.gif`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -73,7 +74,7 @@ describe('Get file info', () => {
 
 	it('PNG with alpha', async (async () => {
 		const path = `${__dirname}/resources/with-alpha.png`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -90,7 +91,7 @@ describe('Get file info', () => {
 
 	it('Generic SVG', async (async () => {
 		const path = `${__dirname}/resources/image.svg`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -108,7 +109,7 @@ describe('Get file info', () => {
 	it('SVG with XML definition', async (async () => {
 		// https://github.com/misskey-dev/misskey/issues/4413
 		const path = `${__dirname}/resources/with-xml-def.svg`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -125,7 +126,7 @@ describe('Get file info', () => {
 
 	it('Dimension limit', async (async () => {
 		const path = `${__dirname}/resources/25000x25000.png`;
-		const info = await getFileInfo(path) as any;
+		const info = await getFileInfoByPath(path) as any;
 		delete info.warnings;
 		delete info.blurhash;
 		assert.deepStrictEqual(info, {
@@ -137,6 +138,23 @@ describe('Get file info', () => {
 			},
 			width: 25000,
 			height: 25000,
+		});
+	}));
+
+	it('Large File', async (async () => {
+		const path = `${__dirname}/resources/P1130536.JPG`;
+		const info = await getFileInfoByPath(path) as any;
+		delete info.warnings;
+		delete info.blurhash;
+		assert.deepStrictEqual(info, {
+			size: 7437824,
+			md5: '7a4eceddbb272aee107507332868519a',
+			type: {
+				mime: 'image/jpeg',
+				ext: 'jpg'
+			},
+			width: 4592,
+			height: 3448,
 		});
 	}));
 });